requirements.txt
troposphere
boto3
nose
pylint
common-fanatics-troposphere